+++ jael/lib/src/ast/identifier.dart
@@ -10,7 +10,7 @@ class Identifier extends Expression {
 
   @override
   compute(SymbolTable scope) {
-    switch(name) {
+    switch (name) {
       case 'null':
         return null;
       case 'true':
@@ -20,14 +20,11 @@ class Identifier extends Expression {
       default:
         var symbol = scope.resolve(name);
         if (symbol == null) {
-          if (scope.resolve('!strict!')?.value == false)
-            return null;
+          if (scope.resolve('!strict!')?.value == false) return null;
           throw new ArgumentError(
               'The name "$name" does not exist in this scope.');
         }
-        return scope
-            .resolve(name)
-            .value;
+        return scope.resolve(name).value;
     }
   }