+++ jael/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.0.3
+* Fix a scanner bug that prevented proper parsing of HTML nodes
+followed by free text.
+* Don't trim `<textarea>` content.
+
 # 1.0.2
 * Use `package:dart2_constant`.
 * Upgrade `package:symbol_table`.